+++ src/main/kotlin/step1/Person.kt
@@ -1,3 +1,5 @@
+package step1
+
 data class Person(
     val name: String,
     val age: Int,
