+++ src/main/kotlin/step2/Operator.kt
@@ -1,3 +1,5 @@
+package step2
+
 enum class Operator(
     private val operator: String
 ) {
