+++ src/test/kotlin/step2/StringCalculatorTest.kt
@@ -1,3 +1,5 @@
+package step2
+
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.Test
